move layouts and error boundaries to lazy
