+++ apps/web/src/components/Layout/IAPLayout.tsx
@@ -25,6 +25,7 @@ import * as micMessages from "~/lang/micCompiled.json";
 
 import SkipLink from "./SkipLink";
 import SitewideBanner from "./SitewideBanner";
+import ErrorBoundary from "./ErrorBoundary/ErrorBoundary";
 
 const messages: Map<string, Messages> = new Map([
   ["crg", crgMessages],
@@ -49,7 +50,7 @@ const IAPSeo = () => {
   );
 };
 
-const Layout = () => {
+export const Component = () => {
   const intl = useIntl();
   const locale = getLocale(intl);
   const location = useLocation();
@@ -103,4 +104,6 @@ const Layout = () => {
   );
 };
 
-export default Layout;
+export { ErrorBoundary };
+
+Component.displayName = "IAPLayout";
